feat: refactor hook imports for consistency and clarity
