+++ src/components/game/components/ui/status/GameStatusBar.tsx
@@ -2,7 +2,7 @@
 
 import type React from "react";
 import { cn } from "@/lib/utils";
-import { useGameStatusBar } from "../../../hooks/GameStatus.hook";
+import { useGameStatusBar } from "../../../hooks/use-game-status";
 
 interface GameStatusBarProps {
   icon: string | React.ReactNode;
