+++ src/components/game/hooks/use-game-status.ts
