+++ src/components/game/hooks/use-tips-popup.ts
